Sample units with proportional to size of the primary unit

Sample units with proportional to size of the primary unit

Usage

pc_pps_survey(data, nsize = 20)

Arguments

data

A data frame object. The data frame must have a column that specifies strata, named stratum, a column that specifies clusters, named cluster, and a column that has size of the cluster, named cluster_pop.

nsize

Number of units to be sampled. If a single number is specified, then equal number of clusters will be sampled from each stratum. If a vector is fed instead, then number of units sampled will depend on the order of the strata in the data frame. To avoid confusions, you must sort your data by stratum before

Value

Returns the the data frame object with additional rows that mark selected units, probability of select, etc

Examples

set.seed(1000)
stratum = rep(c(1,2),c(30,50))
cluster <- c(1:30,1:50)
cluster_pop <- sample(365:1309,80,replace=TRUE)
sample_frame <- data.frame(stratum,cluster, cluster_pop)
pc_pps_survey(data = sample_frame,nsize = 10)
#> $sampled_pps_data
#>    stratum cluster cluster_pop pop_cum_sum sampleSel clusters_sampled
#> 1        1       1         944         944         1               NA
#> 2        1       2         663        1607         2             1238
#> 3        1       3         578        2185         3               NA
#> 4        1       4         415        2600         4               NA
#> 5        1       5         708        3308         5               NA
#> 6        1       6         649        3957         6             3802
#> 7        1       7         719        4676         7               NA
#> 8        1       8        1193        5869         8               NA
#> 9        1       9         766        6635         9             6365
#> 10       1      10         770        7405        10               NA
#> 11       1      11         722        8127        11               NA
#> 12       1      12         665        8792        12               NA
#> 13       1      13         786        9578        13             8929
#> 14       1      14        1293       10871        14               NA
#> 15       1      15        1045       11916        15            11493
#> 16       1      16         905       12821        16               NA
#> 17       1      17        1030       13851        17               NA
#> 18       1      18         806       14657        18            14057
#> 19       1      19         744       15401        19               NA
#> 20       1      20         803       16204        20               NA
#> 21       1      21         894       17098        21            16621
#> 22       1      22         913       18011        22               NA
#> 23       1      23         972       18983        23               NA
#> 24       1      24         458       19441        24            19184
#> 25       1      25         671       20112        25               NA
#> 26       1      26        1308       21420        26               NA
#> 27       1      27        1126       22546        27            21748
#> 28       1      28        1193       23739        28               NA
#> 29       1      29        1206       24945        29            24312
#> 30       1      30         693       25638        30               NA
#> 31       2       1         550         550         1               NA
#> 32       2       2        1220        1770         2             1174
#> 33       2       3        1016        2786         3               NA
#> 34       2       4        1245        4031         4               NA
#> 35       2       5         674        4705         5               NA
#> 36       2       6         516        5221         6               NA
#> 37       2       7         776        5997         7             5463
#> 38       2       8         679        6676         8               NA
#> 39       2       9        1140        7816         9               NA
#> 40       2      10         399        8215        10               NA
#> 41       2      11         903        9118        11               NA
#> 42       2      12         434        9552        12               NA
#> 43       2      13        1011       10563        13             9751
#> 44       2      14         386       10949        14               NA
#> 45       2      15         946       11895        15               NA
#> 46       2      16         819       12714        16               NA
#> 47       2      17         892       13606        17               NA
#> 48       2      18        1032       14638        18            14039
#> 49       2      19         756       15394        19               NA
#> 50       2      20         527       15921        20               NA
#> 51       2      21         769       16690        21               NA
#> 52       2      22        1246       17936        22               NA
#> 53       2      23         985       18921        23            18327
#> 54       2      24         490       19411        24               NA
#> 55       2      25        1278       20689        25               NA
#> 56       2      26        1021       21710        26               NA
#> 57       2      27         861       22571        27               NA
#> 58       2      28         879       23450        28            22615
#> 59       2      29         665       24115        29               NA
#> 60       2      30         730       24845        30               NA
#> 61       2      31        1228       26073        31               NA
#> 62       2      32         916       26989        32            26904
#> 63       2      33         587       27576        33               NA
#> 64       2      34         852       28428        34               NA
#> 65       2      35         567       28995        35               NA
#> 66       2      36        1188       30183        36               NA
#> 67       2      37         967       31150        37               NA
#> 68       2      38        1240       32390        38            31192
#> 69       2      39        1272       33662        39               NA
#> 70       2      40        1094       34756        40               NA
#> 71       2      41         630       35386        41               NA
#> 72       2      42        1047       36433        42            35480
#> 73       2      43         900       37333        43               NA
#> 74       2      44        1062       38395        44               NA
#> 75       2      45         764       39159        45               NA
#> 76       2      46         731       39890        46            39768
#> 77       2      47         664       40554        47               NA
#> 78       2      48         405       40959        48               NA
#> 79       2      49        1090       42049        49               NA
#> 80       2      50         833       42882        50               NA
#>    prob_cluster_sel
#> 1                NA
#> 2         0.2586005
#> 3                NA
#> 4                NA
#> 5                NA
#> 6         0.2531399
#> 7                NA
#> 8                NA
#> 9         0.2987753
#> 10               NA
#> 11               NA
#> 12               NA
#> 13        0.3065762
#> 14               NA
#> 15        0.4075981
#> 16               NA
#> 17               NA
#> 18        0.3143771
#> 19               NA
#> 20               NA
#> 21        0.3487011
#> 22               NA
#> 23               NA
#> 24        0.1786411
#> 25               NA
#> 26               NA
#> 27        0.4391918
#> 28               NA
#> 29        0.4703955
#> 30               NA
#> 31               NA
#> 32        0.2845017
#> 33               NA
#> 34               NA
#> 35               NA
#> 36               NA
#> 37        0.1809617
#> 38               NA
#> 39               NA
#> 40               NA
#> 41               NA
#> 42               NA
#> 43        0.2357633
#> 44               NA
#> 45               NA
#> 46               NA
#> 47               NA
#> 48        0.2406604
#> 49               NA
#> 50               NA
#> 51               NA
#> 52               NA
#> 53        0.2297001
#> 54               NA
#> 55               NA
#> 56               NA
#> 57               NA
#> 58        0.2049811
#> 59               NA
#> 60               NA
#> 61               NA
#> 62        0.2136094
#> 63               NA
#> 64               NA
#> 65               NA
#> 66               NA
#> 67               NA
#> 68        0.2891656
#> 69               NA
#> 70               NA
#> 71               NA
#> 72        0.2441584
#> 73               NA
#> 74               NA
#> 75               NA
#> 76        0.1704678
#> 77               NA
#> 78               NA
#> 79               NA
#> 80               NA
#> 
pc_pps_survey(data = sample_frame,nsize = c(10,7))
#> $sampled_pps_data
#>    stratum cluster cluster_pop pop_cum_sum sampleSel clusters_sampled
#> 1        1       1         944         944         1               NA
#> 2        1       2         663        1607         2             1063
#> 3        1       3         578        2185         3               NA
#> 4        1       4         415        2600         4               NA
#> 5        1       5         708        3308         5               NA
#> 6        1       6         649        3957         6             3627
#> 7        1       7         719        4676         7               NA
#> 8        1       8        1193        5869         8               NA
#> 9        1       9         766        6635         9             6191
#> 10       1      10         770        7405        10               NA
#> 11       1      11         722        8127        11               NA
#> 12       1      12         665        8792        12             8755
#> 13       1      13         786        9578        13               NA
#> 14       1      14        1293       10871        14               NA
#> 15       1      15        1045       11916        15            11318
#> 16       1      16         905       12821        16               NA
#> 17       1      17        1030       13851        17               NA
#> 18       1      18         806       14657        18            13882
#> 19       1      19         744       15401        19               NA
#> 20       1      20         803       16204        20               NA
#> 21       1      21         894       17098        21            16446
#> 22       1      22         913       18011        22               NA
#> 23       1      23         972       18983        23               NA
#> 24       1      24         458       19441        24            19010
#> 25       1      25         671       20112        25               NA
#> 26       1      26        1308       21420        26               NA
#> 27       1      27        1126       22546        27            21574
#> 28       1      28        1193       23739        28               NA
#> 29       1      29        1206       24945        29            24137
#> 30       1      30         693       25638        30               NA
#> 31       2       1         550         550         1               NA
#> 32       2       2        1220        1770         2               NA
#> 33       2       3        1016        2786         3             2673
#> 34       2       4        1245        4031         4               NA
#> 35       2       5         674        4705         5               NA
#> 36       2       6         516        5221         6               NA
#> 37       2       7         776        5997         7               NA
#> 38       2       8         679        6676         8               NA
#> 39       2       9        1140        7816         9               NA
#> 40       2      10         399        8215        10               NA
#> 41       2      11         903        9118        11             8799
#> 42       2      12         434        9552        12               NA
#> 43       2      13        1011       10563        13               NA
#> 44       2      14         386       10949        14               NA
#> 45       2      15         946       11895        15               NA
#> 46       2      16         819       12714        16               NA
#> 47       2      17         892       13606        17               NA
#> 48       2      18        1032       14638        18               NA
#> 49       2      19         756       15394        19            14925
#> 50       2      20         527       15921        20               NA
#> 51       2      21         769       16690        21               NA
#> 52       2      22        1246       17936        22               NA
#> 53       2      23         985       18921        23               NA
#> 54       2      24         490       19411        24               NA
#> 55       2      25        1278       20689        25               NA
#> 56       2      26        1021       21710        26            21051
#> 57       2      27         861       22571        27               NA
#> 58       2      28         879       23450        28               NA
#> 59       2      29         665       24115        29               NA
#> 60       2      30         730       24845        30               NA
#> 61       2      31        1228       26073        31               NA
#> 62       2      32         916       26989        32               NA
#> 63       2      33         587       27576        33            27177
#> 64       2      34         852       28428        34               NA
#> 65       2      35         567       28995        35               NA
#> 66       2      36        1188       30183        36               NA
#> 67       2      37         967       31150        37               NA
#> 68       2      38        1240       32390        38               NA
#> 69       2      39        1272       33662        39            33303
#> 70       2      40        1094       34756        40               NA
#> 71       2      41         630       35386        41               NA
#> 72       2      42        1047       36433        42               NA
#> 73       2      43         900       37333        43               NA
#> 74       2      44        1062       38395        44               NA
#> 75       2      45         764       39159        45               NA
#> 76       2      46         731       39890        46            39429
#> 77       2      47         664       40554        47               NA
#> 78       2      48         405       40959        48               NA
#> 79       2      49        1090       42049        49               NA
#> 80       2      50         833       42882        50               NA
#>    prob_cluster_sel
#> 1                NA
#> 2        0.25860051
#> 3                NA
#> 4                NA
#> 5                NA
#> 6        0.25313987
#> 7                NA
#> 8                NA
#> 9        0.29877526
#> 10               NA
#> 11               NA
#> 12       0.25938061
#> 13               NA
#> 14               NA
#> 15       0.40759810
#> 16               NA
#> 17               NA
#> 18       0.31437710
#> 19               NA
#> 20               NA
#> 21       0.34870115
#> 22               NA
#> 23               NA
#> 24       0.17864108
#> 25               NA
#> 26               NA
#> 27       0.43919182
#> 28               NA
#> 29       0.47039551
#> 30               NA
#> 31               NA
#> 32               NA
#> 33       0.16585047
#> 34               NA
#> 35               NA
#> 36               NA
#> 37               NA
#> 38               NA
#> 39               NA
#> 40               NA
#> 41       0.14740451
#> 42               NA
#> 43               NA
#> 44               NA
#> 45               NA
#> 46               NA
#> 47               NA
#> 48               NA
#> 49       0.12340842
#> 50               NA
#> 51               NA
#> 52               NA
#> 53               NA
#> 54               NA
#> 55               NA
#> 56       0.16666667
#> 57               NA
#> 58               NA
#> 59               NA
#> 60               NA
#> 61               NA
#> 62               NA
#> 63       0.09582109
#> 64               NA
#> 65               NA
#> 66               NA
#> 67               NA
#> 68               NA
#> 69       0.20763957
#> 70               NA
#> 71               NA
#> 72               NA
#> 73               NA
#> 74               NA
#> 75               NA
#> 76       0.11932746
#> 77               NA
#> 78               NA
#> 79               NA
#> 80               NA
#>